Dependencies fix
Former-commit-id: d10ed8356cd0f0f7a54b97c4ccc9ae976580e035
Former-commit-id: 9dcf13c7599f513dde56d6745db14c65d6d2f591 [formerly 084b6103ece3d1c371dda6458553e7e7d951bbe4] [formerly 13e4e3fa1d9c24fc8c80828a6c8d9ad418a4ac08 [formerly 3b6adde123a0d35b980e568fc8facd720b2e810f]]
Former-commit-id: f30c6b94f714e5703189c1af90e00c03ab28a485 [formerly a5237e748f63cbed3e510b2122ea8b176d220ba4]
Former-commit-id: 57ca599a6a97af801344811c8faf81bee634feb8

+++ asciimaker.sh
@@ -172,7 +172,7 @@ time_regex='^([0-5][0-9]:)?[0-5][0-9]:[0-5][0-9]$'
 youtube_dependencies_check(){
   #installs python3, pip3, youtube-dl package and updates pip, setuptools and wheel
   python3_installed=`command -v python3 >/dev/null 2>&1 && echo 1`
-  #pip3_installed=`command -v pip3 >/dev/null 2>&1 && echo 1`
+  pip_installed=`which pip3`
   if [[ ! $python3_installed -eq 1 ]] ; then
     echo -e "${YELLOWBG}'Python 3' is not installed and ASCII Maker needs it to download Youtube videos.${NONE}"
     read -p "Would you like to install it? (Y/N): " confirm && [[ $confirm == [yY] || $confirm == [yY][eE][sS] ]] || exit 1 && clean_temps
@@ -183,11 +183,15 @@ youtube_dependencies_check(){
     fi
     python3 -m pip install --upgrade pip setuptools wheel #check pip, setuptools, and wheel are up to date
   fi
-  #if [[ ! $pip3_installed -eq 1 ]] ; then
-  #  echo -e "${YELLOWBG}'Pip3' is not installed and ASCII Maker needs it to download a Python Youtube package.${NONE}"
-  #  read -p "Would you like to install it? (Y/N): " confirm && [[ $confirm == [yY] || $confirm == [yY][eE][sS] ]] || exit 1 && clean_temps
-  #  sudo apt install python3-pip
-  #fi
+  if [ -z "$pip_installed" ] ; then
+    echo -e "${YELLOWBG}'Pip' is not installed and ASCII Maker needs it to download a Python Youtube package.${NONE}"
+    read -p "Would you like to install it? (Y/N): " confirm && [[ $confirm == [yY] || $confirm == [yY][eE][sS] ]] || exit 1 && clean_temps
+    if [[ "${OSTYPE}" == darwin* ]]; then
+      brew install python3-pip
+    else
+      sudo apt-get install python3-pip
+    fi
+  fi
   if ! python3 -m pip freeze | grep "youtube-dl"= > /dev/null ; then
     echo -e "${YELLOWBG}'youtube-dl' is not installed and ASCII Maker needs it to download Youtube videos.${NONE}"
     read -p "Would you like to install it? (Y/N): " confirm && [[ $confirm == [yY] || $confirm == [yY][eE][sS] ]] || exit 1
